headquarters refactor
